+++ Android/app/src/main/java/com/harvard/studyappmodule/ResourcesWebViewActivity.java
@@ -1,6 +1,6 @@
 /*
  * Copyright © 2017-2019 Harvard Pilgrim Health Care Institute (HPHCI) and its Contributors.
- * Copyright 2020 Google LLC
+ * Copyright 2020-2021 Google LLC
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction, including
  * without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
@@ -11,6 +11,7 @@
  * Funding Source: Food and Drug Administration (“Funding Agency”) effective 18 September 2014 as Contract no. HHSF22320140030I/HHSF22301006T (the “Prime Contract”).
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
  */
 
 package com.harvard.studyappmodule;
@@ -26,42 +27,42 @@ import android.os.Build;
 import android.os.Build.VERSION_CODES;
 import android.os.Bundle;
 import android.os.Environment;
-import android.support.annotation.NonNull;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.FileProvider;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.AppCompatTextView;
 import android.text.Html;
+import android.util.Base64;
 import android.view.View;
 import android.webkit.WebView;
 import android.widget.RelativeLayout;
 import android.widget.Toast;
-import com.github.barteksc.pdfviewer.PDFView;
-import com.github.barteksc.pdfviewer.scroll.DefaultScrollHandle;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.AppCompatTextView;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.FileProvider;
 import com.harvard.R;
+import com.harvard.storagemodule.DbServiceSubscriber;
+import com.harvard.studyappmodule.studymodel.Resource;
 import com.harvard.utils.AppController;
+import com.harvard.utils.CustomFirebaseAnalytics;
 import com.harvard.utils.Logger;
+import com.harvard.utils.PdfViewerView;
 import com.harvard.webservicemodule.apihelper.ConnectionDetector;
-import java.io.BufferedInputStream;
+import io.realm.Realm;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.net.URL;
-import java.net.URLConnection;
 import javax.crypto.CipherInputStream;
+
 
 public class ResourcesWebViewActivity extends AppCompatActivity {
   private AppCompatTextView titleTv;
   private RelativeLayout backBtn;
   private WebView webView;
   private RelativeLayout shareBtn;
-  private PDFView pdfView;
-  private String downloadedFilePath;
+  private String CreateFilePath;
   private String fileName;
-  private String downloadingFileUrl = "";
   private static final int PERMISSION_REQUEST_CODE = 1000;
   private String intentTitle;
   private String intentType;
@@ -69,24 +70,36 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
   private File finalMSharingFile;
   private ConnectionDetector connectionDetector =
       new ConnectionDetector(ResourcesWebViewActivity.this);
+  private Realm realm;
+  private DbServiceSubscriber dbServiceSubscriber;
+  String resourceId;
+  Resource resource;
+  private CustomFirebaseAnalytics analyticsInstance;
+  PdfViewerView pdfViewer;
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
+
     super.onCreate(savedInstanceState);
     setContentView(R.layout.activity_resources_web_view);
+    analyticsInstance = CustomFirebaseAnalytics.getInstance(this);
 
-    downloadedFilePath = "/data/data/" + getPackageName() + "/files/";
+    CreateFilePath = "/data/data/" + getPackageName() + "/files/";
     initializeXmlId();
-    downloadingFileUrl = getIntent().getStringExtra("content");
     intentTitle = getIntent().getStringExtra("title");
     intentType = getIntent().getStringExtra("type");
-    intentContent = getIntent().getStringExtra("content");
+    resourceId = getIntent().getStringExtra("resourceId");
     String studyId = getIntent().getStringExtra("studyId");
+
+    dbServiceSubscriber = new DbServiceSubscriber();
+    realm = AppController.getRealmobj(ResourcesWebViewActivity.this);
+    resource = dbServiceSubscriber.getResource(resourceId, realm);
 
     String title;
     // removing space b/w the string : name of the pdf
     try {
       title = intentTitle.replaceAll("\\s+", "");
+      title = title.replace("/", "\u2215");
     } catch (Exception e) {
       title = intentTitle;
       Logger.log(e);
@@ -113,7 +126,7 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
         } else {
           if (connectionDetector.isConnectingToInternet()) {
             // starting new Async Task for downlaoding pdf file
-            new DownloadFileFromUrl(downloadingFileUrl, downloadedFilePath, fileName).execute();
+            new CreateFileFromBase64(resource.getContent(), CreateFilePath, fileName).execute();
           } else {
             // offline functionality
             offLineFunctionality();
@@ -122,7 +135,7 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
       } else {
         if (connectionDetector.isConnectingToInternet()) {
           // starting new Async Task for downlaoding pdf file
-          new DownloadFileFromUrl(downloadingFileUrl, downloadedFilePath, fileName).execute();
+          new CreateFileFromBase64(resource.getContent(), CreateFilePath, fileName).execute();
         } else {
           // offline functionality
           offLineFunctionality();
@@ -138,6 +151,12 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
         new View.OnClickListener() {
           @Override
           public void onClick(View view) {
+            Bundle eventProperties = new Bundle();
+            eventProperties.putString(
+                CustomFirebaseAnalytics.Param.BUTTON_CLICK_REASON,
+                getString(R.string.resources_webview_back));
+            analyticsInstance.logEvent(
+                CustomFirebaseAnalytics.Event.ADD_BUTTON_CLICK, eventProperties);
             finish();
           }
         });
@@ -145,6 +164,12 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
         new View.OnClickListener() {
           @Override
           public void onClick(View v) {
+            Bundle eventProperties = new Bundle();
+            eventProperties.putString(
+                CustomFirebaseAnalytics.Param.BUTTON_CLICK_REASON,
+                getString(R.string.resources_webview_share));
+            analyticsInstance.logEvent(
+                CustomFirebaseAnalytics.Event.ADD_BUTTON_CLICK, eventProperties);
             try {
 
               Intent shareIntent = new Intent(Intent.ACTION_SEND);
@@ -154,7 +179,7 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
               shareIntent.putExtra(Intent.EXTRA_SUBJECT, intentTitle);
               // if pdf then attach and send else content send
               if (intentType.equalsIgnoreCase("pdf")) {
-                File file = new File(downloadedFilePath + fileName + ".pdf");
+                File file = new File(CreateFilePath + fileName + ".pdf");
                 if (file.exists()) {
                   finalMSharingFile = copy(file);
                   Uri fileUri =
@@ -165,9 +190,12 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
                   shareIntent.putExtra(Intent.EXTRA_STREAM, fileUri);
                 }
               } else {
-                shareIntent.putExtra(Intent.EXTRA_TEXT, Html.fromHtml(intentContent));
+                shareIntent.setType("text/html");
+                shareIntent.putExtra(
+                    Intent.EXTRA_TEXT,Html.fromHtml(resource.getContent()).toString());
+                shareIntent.putExtra(
+                    Intent.EXTRA_HTML_TEXT, Html.fromHtml(resource.getContent()).toString());
               }
-
               startActivity(shareIntent);
             } catch (Exception e) {
               Logger.log(e);
@@ -189,7 +217,7 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
         finish();
       } else {
         if (connectionDetector.isConnectingToInternet()) {
-          new DownloadFileFromUrl(downloadingFileUrl, downloadedFilePath, fileName).execute();
+          new CreateFileFromBase64(resource.getContent(), CreateFilePath, fileName).execute();
         } else {
           Toast.makeText(
                   ResourcesWebViewActivity.this,
@@ -206,7 +234,7 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
     titleTv = (AppCompatTextView) findViewById(R.id.title);
     webView = (WebView) findViewById(R.id.webView);
     shareBtn = (RelativeLayout) findViewById(R.id.shareBtn);
-    pdfView = (PDFView) findViewById(R.id.pdfView);
+    pdfViewer = (PdfViewerView) findViewById(R.id.pdfViewer);
   }
 
   private void setTextForView() {
@@ -216,12 +244,17 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
     webView.getSettings().setJavaScriptEnabled(true);
     webView.getSettings().setDefaultTextEncodingName("utf-8");
     webView.setScrollBarStyle(View.SCROLLBARS_INSIDE_OVERLAY);
-    String webData = intentContent;
+    String webData = resource.getContent();
     if (Build.VERSION.SDK_INT >= 24) {
-      webView.loadData(
-              Html.fromHtml((webData), Html.FROM_HTML_MODE_LEGACY).toString(), "text/html", "UTF-8");
+      webView.loadDataWithBaseURL(
+          null,
+          Html.fromHtml((webData), Html.FROM_HTML_MODE_LEGACY).toString(),
+          "text/html",
+          "UTF-8",
+          null);
     } else {
-      webView.loadData(Html.fromHtml((webData)).toString(), "text/html", "UTF-8");
+      webView.loadDataWithBaseURL(
+          null, Html.fromHtml((webData)).toString(), "text/html", "UTF-8", null);
     }
   }
 
@@ -272,7 +305,7 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
   protected void onDestroy() {
     super.onDestroy();
     try {
-      File file = new File(downloadedFilePath + fileName + ".pdf");
+      File file = new File(CreateFilePath + fileName + ".pdf");
       if (file.exists()) {
         file.delete();
       }
@@ -280,24 +313,27 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
       Logger.log(e);
     }
     try {
-      if (finalMSharingFile.exists()) {
+      if (finalMSharingFile != null && finalMSharingFile.exists()) {
         finalMSharingFile.delete();
       }
 
     } catch (Exception e) {
       Logger.log(e);
     }
+    dbServiceSubscriber.closeRealmObj(realm);
   }
 
-  class DownloadFileFromUrl extends AsyncTask<String, String, String> {
+  class CreateFileFromBase64 extends AsyncTask<String, String, String> {
 
-    /** Before starting background thread Show Progress Bar Dialog. */
+    /**
+     * Before starting background thread Show Progress Bar Dialog.
+     */
     String downloadUrl = "";
 
     String filePath = "";
     String fileName = "";
 
-    DownloadFileFromUrl(String downloadUrl, String filePath, String fileName) {
+    CreateFileFromBase64(String downloadUrl, String filePath, String fileName) {
       this.downloadUrl = downloadUrl;
       this.filePath = filePath;
       this.fileName = fileName;
@@ -307,46 +343,19 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
     protected void onPreExecute() {
       super.onPreExecute();
       AppController.getHelperProgressDialog()
-          .showProgress(ResourcesWebViewActivity.this, "", "", false);
+              .showProgress(ResourcesWebViewActivity.this, "", "", false);
     }
 
-    /** Downloading file in background thread. */
+    /**
+     * Downloading file in background thread.
+     */
     @Override
     protected String doInBackground(String... url1) {
       int count;
       try {
-        URL url = new URL(downloadUrl);
-        URLConnection conection = url.openConnection();
-        conection.connect();
-        // getting file length
-        int lenghtOfFile = conection.getContentLength();
-
-        // input stream to read file - with 8k buffer
-        InputStream input = new BufferedInputStream(url.openStream(), 8192);
-
-        // Output stream to write file
-        OutputStream output = new FileOutputStream(filePath + fileName + ".pdf");
-
-        byte[] data = new byte[1024];
-
-        long total = 0;
-
-        while ((count = input.read(data)) != -1) {
-          total += count;
-          // publishing the progress....
-          // After this onProgressUpdate will be called
-          publishProgress("" + (int) ((total * 100) / lenghtOfFile));
-
-          // writing data to file
-          output.write(data, 0, count);
-        }
-
-        // flushing output
-        output.flush();
-
-        // closing streams
-        output.close();
-        input.close();
+        FileOutputStream fos = new FileOutputStream(filePath + fileName + ".pdf");
+        fos.write(Base64.decode(downloadUrl.split(",")[1], Base64.NO_WRAP));
+        fos.close();
 
       } catch (Exception e) {
         // while downloading time, net got disconnected so delete the file
@@ -356,17 +365,19 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
           Logger.log(e1);
         }
       }
+      AppController.generateEncryptedConsentPdf(filePath, fileName);
       return null;
     }
 
-    /** After completing background task Dismiss the progress dialog. */
+    /**
+     * After completing background task Dismiss the progress dialog.
+     */
     @Override
     protected void onPostExecute(String url) {
       try {
         // downlaod success mean file exist else check offline file
         File file = new File(filePath + fileName + ".pdf");
         if (file.exists()) {
-          AppController.generateEncryptedConsentPdf(filePath, fileName);
           displayPdfView(filePath + fileName + ".pdf");
         } else {
           // offline functionality
@@ -383,10 +394,10 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
   private void offLineFunctionality() {
     try {
       // checking encrypted file is there or not?
-      File file = new File(downloadedFilePath + fileName + ".txt");
+      File file = new File(CreateFilePath + fileName + ".txt");
       if (file.exists()) {
         // decrypt the file
-        File decryptFile = getEncryptedFilePath(downloadedFilePath + fileName + ".txt");
+        File decryptFile = getEncryptedFilePath(CreateFilePath + fileName + ".txt");
         displayPdfView(decryptFile.getAbsolutePath());
       } else {
         Toast.makeText(
@@ -409,7 +420,7 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
     try {
       CipherInputStream cis = AppController.generateDecryptedConsentPdf(filePath);
       byte[] byteArray = AppController.cipherInputStreamConvertToByte(cis);
-      File file = new File(downloadedFilePath + fileName + ".pdf");
+      File file = new File(CreateFilePath + fileName + ".pdf");
       if (!file.exists() && file == null) {
         file.createNewFile();
       }
@@ -424,16 +435,7 @@ public class ResourcesWebViewActivity extends AppCompatActivity {
   }
 
   private void displayPdfView(String filePath) {
-    pdfView.setVisibility(View.VISIBLE);
-    try {
-      pdfView
-          .fromFile(new File(filePath))
-          .defaultPage(0)
-          .enableAnnotationRendering(true)
-          .scrollHandle(new DefaultScrollHandle(ResourcesWebViewActivity.this))
-          .load();
-    } catch (Exception e) {
-      Logger.log(e);
-    }
+    pdfViewer.setVisibility(View.VISIBLE);
+    pdfViewer.setPdf(new File(filePath));
   }
 }
